Update README.md
Adding Organization Members to a Team:
People with owner or team maintainer permissions can add organization members to teams.
In the upper-right corner of GitHub, select your profile photo, then click Your organizations.
Click the name of your organization.
Under your organization name, click Teams.
Click the name of the team.
At the top of the team page, click Members.
Above the list of team members, click Add a member.
Add an organization member to the team or invite a non-member to join a team by typing their username or email address and clicking Enter. The user will receive an email invitation to join your organization and will be added to the team once they accept the invitation1.
Managing Teams and People with Access to Your Repository:
Navigate to the main page of the repository.
Under your repository name, click Settings.
In the Access section of the sidebar, click Collaborators & teams.
To the right of Manage access, click Add people or Add teams.
In the search field, start typing the name of the team or person to invite, then click a name in the list of matches.
Under Choose a role, select the repository role to grant to the team or person, then click Add NAME to REPOSITORY2.